src/roadmap-gen/requirements.txt
asgi_correlation_id==4.3.4
fastapi==0.104.0
openai==1.64.0
pydantic==2.4.2
pydantic-settings
python-dotenv==1.0.1
python-multipart==0.0.9
SQLAlchemy==2.0.38
streamlit==1.39.0
structlog==22.3.0
uvicorn[standard]==0.18.3
